feat(TDD cycle: Red stage): Add failing test for `withdraw` method.
The new test describe a desired behavior for the `Account` class:
"It should withdraw an amount of money from the account by updating
its balance. The withdrawal amount must not exceed the balance".

This test "should fail!", because the code to make it pass hasn't been
written yet (`withdraw` method doesn't exist yet!).